refactor: extract /admin/internals routes into ui-routes module (A2)
Move the four system-internals / threads-tools routes verbatim out of
lib/routes-ui.js into lib/ui-routes/internals-routes.js:
  GET  /admin/internals
  POST /admin/internals/kill
  POST /admin/internals/snapshot
  GET  /admin/internals/thread/{threadId}

formatAccountData is imported from the shared route-helpers module. Wiring
mirrors the existing extracted modules (init({ server, call }) in applyRoutes).

routes-ui.js: 8143 -> 7813 lines; size ratchet lowered accordingly. Route table
verified unchanged via the snapshot test; GET pages reachable (200) on a booted
server; no new security surface (verbatim move, validation preserved).

Stacked on the Milestone 1 foundation branch.
